[SaiKavya] adding service interface

+++ MongoDBCRUD/src/main/java/com/bridgelabz/mongo/serviceprovider/UserService.java
@@ -6,8 +6,8 @@ import com.bridgelabz.mongo.model.UserModel;
 import com.mongodb.client.MongoCollection;
 
 public interface UserService {
-	public void insertIntoMongoDB(MongoCollection<UserModel> collection, int id, String userName, String email, long phoneNumber, String city);
+	public void insertIntoMongoDB(MongoCollection<UserModel> collection, long id, String userName, String email, long phoneNumber, String city);
 	public List<UserModel> readFromMongoDB(MongoCollection<UserModel> userModel);
-	public void updateMongoDB(MongoCollection<UserModel> collection, String newUserName, int id);
+	public void updateMongoDB(MongoCollection<UserModel> collection, String newUserName, long id);
 	public void deleteDetails(MongoCollection<UserModel> collection, String userName);
 }
